requirements.txt
yt-dlp>=2024.12.23
youtube-transcript-api==0.6.2
google-api-python-client==2.149.0
anthropic>=0.40.0
python-dotenv==1.0.1
pyyaml>=6.0
